Add swagger

+++ ppac-service-impl/src/main/kotlin/com/levi9/ppac/service/api/service/CompanyServiceImpl.kt
@@ -1,25 +1,38 @@
 package com.levi9.ppac.service.api.service
 
 import com.levi9.ppac.service.api.data_classes.Company
+import com.levi9.ppac.service.api.repository.CodeRepository
 import com.levi9.ppac.service.api.repository.CompanyRepository
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.data.repository.findByIdOrNull
+import org.springframework.http.HttpStatus
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
+import org.springframework.web.server.ResponseStatusException
 import java.util.*
 
 @Service
 @ConditionalOnProperty(prefix = "feature", name = ["mvp"], havingValue = "true")
 class CompanyServiceImpl(
-    val companyRepository: CompanyRepository
+    val companyRepository: CompanyRepository,
+    val codeRepository: CodeRepository
 ) : CompanyService<Company> {
     @Transactional
     override fun findAll(adminCode: Int): List<Company> {
+
+        if (!codeRepository.isAdminCode(adminCode)) {
+            throw ResponseStatusException(HttpStatus.UNAUTHORIZED)
+        }
+
         return companyRepository.findAll().map { Company.parse(it) }
     }
 
     @Transactional
     override fun deleteById(adminCode: Int, id: UUID) {
+
+        if (!codeRepository.isAdminCode(adminCode)) {
+            throw ResponseStatusException(HttpStatus.UNAUTHORIZED)
+        }
 
         companyRepository.findByIdOrNull(id)?.let {
             companyRepository.deleteById(id)
